✨ feat [#10.11.10]: Metadata API 다국어화 및 라우팅 구조 최적화
🔧 변경 사항:
- **[Structure]**: [app]
  - `graph`, `stats` 페이지를 `[locale]` 하위로 이동하여 다국어 라우팅 적용
  - Root Layout(`app/layout.tsx`)을 `app/[locale]/layout.tsx`로 통합 및 마이그레이션
- **[SEO]**: [generateMetadata]
  - 각 페이지(Dashboard, Graph, Stats)별 동적 메타데이터(Title) 구현
  - [layout.tsx](web_ui/src/app/layout.tsx)에 기본 메타데이터 템플릿(`%s | FlowNote`) 적용
  - Next.js 15+ 호환성을 위해 `await params` 패턴 적용
- **[Locales]**: [en.json, ko.json]
  - `metadata` 네임스페이스 추가

🎯 목적:
- SEO 최적화 및 페이지별 다국어 타이틀 지원
- 올바른 i18n 라우팅 구조 확립

📌 Related:
- Issue [#275]
