+++ web_ui/src/app/[locale]/layout.tsx
@@ -2,20 +2,52 @@
 
 import { NextIntlClientProvider } from 'next-intl';
 import { notFound } from 'next/navigation';
+import { getTranslations } from 'next-intl/server';
 import { locales, type Locale } from '@/i18n/config';
+import { Sidebar } from "@/components/layout/sidebar";
+import { MobileNav } from "@/components/layout/mobile-nav";
+import { Toaster } from "@/components/ui/sonner";
+import { Geist, Geist_Mono } from "next/font/google";
+import "../globals.css";
+
+const geistSans = Geist({
+  variable: "--font-geist-sans",
+  subsets: ["latin"],
+});
+
+const geistMono = Geist_Mono({
+  variable: "--font-geist-mono",
+  subsets: ["latin"],
+});
 
 export function generateStaticParams() {
   return locales.map((locale) => ({ locale }));
 }
 
+export async function generateMetadata({ params }: { params: Promise<{ locale: Locale }> }) {
+  const { locale } = await params;
+  const t = await getTranslations({ locale, namespace: 'metadata' });
+  return {
+    title: {
+      default: t('title'),
+      template: `%s | ${t('title')}`,
+    },
+    description: t('description'),
+    icons: {
+      icon: "/favicon.ico",
+    },
+  };
+}
+
 export default async function LocaleLayout({
   children,
-  params: { locale }
+  params
 }: {
   children: React.ReactNode;
-  params: { locale: Locale };
+  params: Promise<{ locale: Locale }>;
 }) {
-  // 지원하지 않는 locale인 경우 404
+  const { locale } = await params;
+  
   if (!locales.includes(locale)) {
     notFound();
   }
@@ -28,8 +60,19 @@ export default async function LocaleLayout({
   }
 
   return (
-    <NextIntlClientProvider locale={locale} messages={messages}>
-      {children}
-    </NextIntlClientProvider>
+    <html lang={locale}>
+      <body className={`${geistSans.variable} ${geistMono.variable} antialiased bg-slate-50`}>
+        <NextIntlClientProvider locale={locale} messages={messages}>
+          <Sidebar className="hidden md:flex" />
+          <div className="flex flex-col min-h-screen md:pl-64 transition-all duration-300 ease-in-out">
+            <MobileNav />
+            <main className="flex-1 p-4 md:p-8 pt-6">
+              {children}
+            </main>
+          </div>
+          <Toaster />
+        </NextIntlClientProvider>
+      </body>
+    </html>
   );
 }
