+++ web_ui/src/app/[locale]/loading.tsx
