+++ web_ui/src/locales/ko.json
@@ -75,5 +75,9 @@
         "unknown": "알 수 없는 전략"
       }
     }
+  },
+  "metadata": {
+    "title": "FlowNote - AI 지식 관리 시스템",
+    "description": "PARA 방법론 기반의 AI 지식 관리 시스템. 당신의 제2의 두뇌를 자동화하세요."
   }
 }
